% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_list.R
\name{bucket_list_survey}
\alias{bucket_list_survey}
\title{Create a bucket list for use with surveydown}
\usage{
bucket_list_survey(
  header = NULL,
  ...,
  group_name,
  css_id = group_name,
  group_put_max = rep(Inf, length(labels)),
  options = sortable_options(),
  class = "default-sortable",
  orientation = c("horizontal", "vertical"),
  output_width = "100\%",
  output_height = "100\%"
)
}
\arguments{
\item{header}{Text that appears at the top of the bucket list.  (This is
encoded as an HTML \verb{<p>} tag, so not strictly speaking a header.)  Note
that you must explicitly provide \code{header} argument, especially in the case
where you want the header to be empty - to do this use \code{header = NULL} or
\code{header = NA}.}

\item{...}{One or more specifications for a rank list, and must be defined by
\link{add_rank_list}.}

\item{group_name}{Passed to \code{SortableJS} as the group name. Also the input
value set in Shiny. (\code{input[[group_name]]}). Items can be dragged between
bucket lists which share the same group name.}

\item{css_id}{This is the css id to use, and must be unique in your shiny
app. This defaults to the value of \code{group_id}, and will be appended to the
value "bucket-list-container", to ensure the CSS id is unique for the
container as well as the embedded rank lists.}

\item{group_put_max}{Not yet implemented}

\item{options}{Options to be supplied to \link{sortable_js} object. See \link{sortable_options} for more details}

\item{class}{A css class applied to the bucket list and rank lists.  This can
be used to define custom styling.}

\item{orientation}{Either \code{horizontal} or \code{vertical}, and specifies the
layout of the components on the page.}

\item{output_width}{The width of the Shiny widget. Necessary for use
\link{pkg:surveydown}.}

\item{output_height}{The height of the Shiny widget. Necessary for use
\link{pkg:surveydown}.}
}
\value{
A list with class \code{bucket_list}
}
\description{
This is a clone of the \code{\link[=bucket_list]{bucket_list()}} question that is compatible
with the \link{pkg:surveydown} Quarto Shiny interactive format.
All options are the same as the standard \code{\link[=bucket_list]{bucket_list()}} question.
}
\seealso{
\link{rank_list_survey}, \link{update_rank_list}
}
