% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_list.R
\name{rank_list_survey}
\alias{rank_list_survey}
\title{Ranking question for use in surveydown}
\usage{
rank_list_survey(
  text = "",
  labels,
  input_id,
  css_id = input_id,
  options = sortable_options(),
  orientation = c("vertical", "horizontal"),
  class = "default-sortable",
  output_width = "100\%",
  output_height = "100\%"
)
}
\arguments{
\item{text}{Text to appear at top of list.}

\item{labels}{A character vector with the text to display inside the widget.
This can also be a list of html tag elements.  The text content of each
label or label name will be used to set the shiny \code{input_id} value.
To create an empty \code{rank_list}, use \code{labels = list()}.}

\item{input_id}{output variable to read the plot/image from.}

\item{css_id}{This is the css id to use, and must be unique in your shiny
app. This defaults to the value of \code{input_id}, and will be appended to the
value "rank-list-container", to ensure the CSS id is unique for the
container as well as the labels.
If NULL, the function generates an id of the form
\code{rank_list_id_1}, and will automatically increment for every \code{rank_list}.}

\item{options}{Options to be supplied to \link{sortable_js} object. See \link{sortable_options} for more details}

\item{orientation}{Set this to "horizontal" to get horizontal orientation of
the items.}

\item{class}{A css class applied to the rank list.  This can be used to
define custom styling.}

\item{output_width}{Width of the html widget. Need to set for
compatibility with Quarto and surveydown.}

\item{output_height}{Height of the html widget. Need to set for
compatibility with Quarto and surveydown.}
}
\description{
This question is a clone of \link{rank_list} but with
some modifications to allow it to be used with in the \link{surveydown}
R package. For more information about function options,
see \link{rank_list} documentation.
}
